refactor(ui): extract TestListSearchMixin from run explorer (#548)
The fuzzy search box + goto (/, Esc, live filter, highlight-best, commit)
moves into a reusable DOMNode mixin parameterized by an abstract
_compute_options and optional _extra_predicate / _extra_filter_labels
hooks. RunTestExplorerScreen keeps only its failing-only specifics. No
behavior change (existing tests green); sets up reuse by the build
test explorer.
